marked up the top page by using haml
index.html.haml is rendering two partial templates. These templates
should be rendered by using each method so will be revised in further
step but not in this branch.

1, _group.html.haml
Partial template for the group names in the left column of the page.
This file should not be located in messages directory so will be
located in appropriate dir in further steps but not in this branch.

2, _message.html.haml
Partial template for the messages in the right column of the page.

Also, form tag should be replaced by form_for helper method so will be
revised after generating model in further step but not in this branch.